[Decomp][Player] Cohesion split: extract pvp methods to PlayerPvP.cpp
